% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{analyzeGrowthRate.mwIPM}
\alias{analyzeGrowthRate.mwIPM}
\title{Compute the population growth rate.}
\usage{
\method{analyzeGrowthRate}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
The population growth rate
}
\description{
Compute the population growth rate.
}
\examples{
lambda <- ipm \%>\% analyzeGrowthRate()
}
