% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{analyzeParameters}
\alias{analyzeParameters}
\title{Sensitivity analysis on parameters.}
\usage{
analyzeParameters(obj, compute, saveresults, params, distpars)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{compute}{Recompute or load from cache?}

\item{saveresults}{Cache results?}

\item{params}{Which parameter to perturb?  One of "all", "Flowering", ...}

\item{distpars}{Determines in which scale (log or linear) perturbation is occurring.}
}
\value{
A mwIPM model object, with the results of the computation stored
in a list.
}
\description{
Sensitivity analysis on parameters.
}
\examples{
ipm \%<>\% analyzeParameters()
str(ipm$analysis$parameters)
}
