% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{bootIPM}
\alias{bootIPM}
\title{Bootstraps over the stems.}
\usage{
bootIPM(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A mwIPM model object.
}
\description{
Bootstraps over the stems.
}
\examples{
ipm \%<>\% bootIPM()
}
