% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_mod.R
\name{calcPars}
\alias{calcPars}
\title{Store scaled and calculate unscaled parameters for models.}
\usage{
calcPars(mdl, scaled, vars)
}
\arguments{
\item{mdl}{Model}

\item{scaled}{List of scaled variables (output of `scale` command)}

\item{vars}{Vector of names for scaled variables}
}
\value{
List of scaled and unscaled parameters.
}
\description{
Store scaled and calculate unscaled parameters for models.
}
