% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_mod.R
\name{checkPars}
\alias{checkPars}
\title{Check to make sure parameter unscaling worked after call to
\code{\link{calcPars}} function.}
\usage{
checkPars(obj)
}
\arguments{
\item{obj}{mwMod model object}
}
\description{
Check to make sure parameter unscaling worked after call to
\code{\link{calcPars}} function.
}
