% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{computeFullKernel.mwIPM}
\alias{computeFullKernel.mwIPM}
\title{Compute full kernel.}
\usage{
\method{computeFullKernel}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A mwIPM model object.
}
\description{
Compute full kernel.
}
