% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{computeMPM.mwIPM}
\alias{computeMPM.mwIPM}
\title{Compute the MPM from the IPM}
\usage{
\method{computeMPM}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A mwIPM model object.
}
\description{
Compute the MPM from the IPM
}
\examples{
ipm \%<>\% computeMPM()
}
