% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{computeSexualKernel.mwIPM}
\alias{computeSexualKernel.mwIPM}
\title{Compute sexual reproduction kernel.}
\usage{
\method{computeSexualKernel}{mwIPM}(obj, update = TRUE, perturb = rep(0, 2))
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{update}{Update dependencies?}

\item{perturb}{Parameter perturbation vector for sensitivity analysis.}
}
\value{
A mwIPM model object.
}
\description{
Compute sexual reproduction kernel.
}
