% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{jacMomsToPars}
\alias{jacMomsToPars}
\title{Jacobian of moments to parameters.}
\usage{
jacMomsToPars(y, type = "ident")
}
\arguments{
\item{y}{Vector of moments.}

\item{type}{Distribution type (lnorm, gamma, or identity)}
}
\value{
Jacobian matrix
}
\description{
Jacobian of moments to parameters.
}
