% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{MomsToPars}
\alias{MomsToPars}
\title{Moments to parameters.}
\usage{
MomsToPars(y, type = "lnorm")
}
\arguments{
\item{y}{Vector of moments.}

\item{type}{Distribution type (lnorm, gamma, or identity)}
}
\value{
Vector of moments.
}
\description{
Moments to parameters.
}
