% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{mwIPM}
\alias{mwIPM}
\title{Construcor method for mwIPM S3 class.}
\usage{
mwIPM(x = list())
}
\arguments{
\item{x}{A list of options.}
}
\value{
A mwIPM model object.
}
\description{
Construcor method for mwIPM S3 class.
}
\examples{
ipm <- mwIPM()
}
