% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{ParsToMoms}
\alias{ParsToMoms}
\title{Parameters to moments.}
\usage{
ParsToMoms(x, type = "lnorm")
}
\arguments{
\item{x}{Vector of parameters.}

\item{type}{Distribution type (lnorm, gamma, or identity)}
}
\value{
Vector of moments.
}
\description{
Parameters to moments.
}
