% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{perturbTrans}
\alias{perturbTrans}
\title{Perturbation of the parameter-to-moments transformation.}
\usage{
perturbTrans(pars, perturb = rep(0, 2), type = "lnorm")
}
\arguments{
\item{pars}{Vector of parameters or moments}

\item{perturb}{Perturbation vector}

\item{type}{Distribution type (lnorm, gamma, or identity)}
}
\value{
Gradient vector.
}
\description{
Perturbation of the parameter-to-moments transformation.
}
