% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_mod.R
\name{predict.mwMod}
\alias{predict.mwMod}
\title{Method for mwMod prediction}
\usage{
\method{predict}{mwMod}(obj, newdata, type = "Bertha", perturb = rep(0,
  5))
}
\arguments{
\item{obj}{Object of class mwMod.}

\item{newdata}{Data frame with new data for prediction.}

\item{type}{Site}

\item{perturb}{Perturbation vector for sensitivity analysis.}
}
\description{
Method for mwMod prediction
}
