% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_mod.R
\name{predict}
\alias{predict}
\title{Method for mwMod prediction}
\usage{
predict(obj, newdata, type, perturb)
}
\arguments{
\item{obj}{Object of class mwMod.}

\item{newdata}{Data frame with new data for prediction.}

\item{type}{Site}

\item{perturb}{Perturbation vector for sensitivity analysis.}
}
\description{
Method for mwMod prediction
}
