% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{renderBudlingDistFit.mwIPM}
\alias{renderBudlingDistFit.mwIPM}
\title{Render plot of budling distributions.}
\usage{
\method{renderBudlingDistFit}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A plot object.
}
\description{
Render plot of budling distributions.
}
\examples{
ipm \%>\% renderBudlingDistFit()
}
