% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{renderFloweringFit.mwIPM}
\alias{renderFloweringFit.mwIPM}
\title{Render plot of flowering vs. height.}
\usage{
\method{renderFloweringFit}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A plot object.
}
\description{
Render plot of flowering vs. height.
}
\examples{
ipm \%>\% renderFlowerFit()
}
