% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{renderHerbivoryDistFit.mwIPM}
\alias{renderHerbivoryDistFit.mwIPM}
\title{Render plot of herbivory distributions.}
\usage{
\method{renderHerbivoryDistFit}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A plot object.
}
\description{
Render plot of herbivory distributions.
}
\examples{
ipm \%>\% renderHerbivoryDistFit()
}
