% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{requirePackages}
\alias{requirePackages}
\title{Helper function to check for required packages.}
\usage{
requirePackages(req)
}
\arguments{
\item{req}{Vector of required packages.}
}
\description{
Helper function to check for required packages.
}
