% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setGrowthMatrix.mwIPM}
\alias{setGrowthMatrix.mwIPM}
\title{Create growth matrix.}
\usage{
\method{setGrowthMatrix}{mwIPM}(obj, update = TRUE, perturb = rep(0,
  5))
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{update}{Update dependencies?}

\item{perturb}{Parameter perturbation vector for sensitivity analysis.}
}
\value{
A mwIPM model object.
}
\description{
Create growth matrix.
}
