% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setHerbivoryDistFit.mwIPM}
\alias{setHerbivoryDistFit.mwIPM}
\title{Sets the herbivory distribution.}
\usage{
\method{setHerbivoryDistFit}{mwIPM}(obj, compute = FALSE,
  saveresults = FALSE, update = TRUE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{compute}{Recompute or load from cache?}

\item{saveresults}{Cache results?}

\item{update}{Update dependencies?}
}
\value{
A mwIPM model object.
}
\description{
Sets the herbivory distribution.
}
