% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setHerbivoryMatrix.mwIPM}
\alias{setHerbivoryMatrix.mwIPM}
\title{Create herbivory matrix.}
\usage{
\method{setHerbivoryMatrix}{mwIPM}(obj, dist.herb = NA, update = TRUE,
  perturb = rep(0, 3), distpars = FALSE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{dist.herb}{Herbivory distribution.}

\item{update}{Update dependencies?}

\item{perturb}{Parameter perturbation vector for sensitivity analysis.}

\item{distpars}{Determines in which scale (log or linear) perturbation is occurring.}
}
\value{
A mwIPM model object.
}
\description{
Create herbivory matrix.
}
