% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setModel.mwIPM}
\alias{setModel.mwIPM}
\title{Set site and recompute kernel.}
\usage{
\method{setModel}{mwIPM}(obj, model = "Bertha", compute = FALSE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{model}{Model "Bertha", site, or year}

\item{compute}{Recompute or load from cache?}
}
\value{
A mwIPM model object.
}
\description{
Set site and recompute kernel.
}
\examples{
ipm \%<>\% setModel("BLD1")
}
