% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setModel}
\alias{setModel}
\title{Set model and recompute kernel.}
\usage{
setModel(obj, model, compute)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{model}{Model "Bertha", site, or year}

\item{compute}{Recompute or load from cache?}
}
\value{
A mwIPM model object.
}
\description{
Set model and recompute kernel.
}
\examples{
ipm \%<>\% setModel("BLD1")
}
