% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setPodsFit.mwIPM}
\alias{setPodsFit.mwIPM}
\title{Sets the pods mixed-model.}
\usage{
\method{setPodsFit}{mwIPM}(obj, compute = FALSE, saveresults = FALSE, update = TRUE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{compute}{Recompute or load from cache?}

\item{saveresults}{Cache results?}

\item{update}{Update dependencies?}
}
\value{
A mwIPM model object.
}
\description{
Sets the pods mixed-model.
}
