% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setPodsMatrix.mwIPM}
\alias{setPodsMatrix.mwIPM}
\title{Create pods matrix.}
\usage{
\method{setPodsMatrix}{mwIPM}(obj, compute = FALSE, saveresults = FALSE,
  update = TRUE, perturb = rep(0, 5))
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{update}{Update dependencies?}

\item{perturb}{Parameter perturbation vector for sensitivity analysis.}
}
\value{
A mwIPM model object.
}
\description{
Create pods matrix.
}
