% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setSeedlingDistFit.mwIPM}
\alias{setSeedlingDistFit.mwIPM}
\title{Sets the seedling distribution (normal distribution).}
\usage{
\method{setSeedlingDistFit}{mwIPM}(obj, compute = FALSE, saveresults = FALSE, update = TRUE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{compute}{Recompute or load from cache?}

\item{saveresults}{Cache results?}

\item{update}{Update dependencies?}
}
\value{
A mwIPM model object.
}
\description{
Sets the seedling distribution (normal distribution).
}
