% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setSeedlingRecruitmentMatrix.mwIPM}
\alias{setSeedlingRecruitmentMatrix.mwIPM}
\title{Create seedling recruitment matrix.}
\usage{
\method{setSeedlingRecruitmentMatrix}{mwIPM}(obj, update = TRUE, perturb = rep(0, 2))
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{update}{Update dependencies?}

\item{perturb}{Parameter perturbation vector for sensitivity analysis.}
}
\value{
A mwIPM model object.
}
\description{
Create seedling recruitment matrix.
}
