% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setSeedsPerPodConst.mwIPM}
\alias{setSeedsPerPodConst.mwIPM}
\title{Sets the number of seeds per pod parameter.}
\usage{
\method{setSeedsPerPodConst}{mwIPM}(obj, compute = FALSE,
  saveresults = FALSE, update = TRUE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{compute}{Recompute or load from cache?}

\item{saveresults}{Cache results?}

\item{update}{Update dependencies?}
}
\value{
A mwIPM model object.
}
\description{
Sets the number of seeds per pod parameter.
}
