% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setSite.mwIPM}
\alias{setSite.mwIPM}
\title{Set site and recompute kernel.}
\usage{
\method{setSite}{mwIPM}(obj, site = "Bertha", compute = FALSE)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{site}{Site ("Bertha", "BLD1", "BLD2", "PWR", "SKY", "YTB")}

\item{compute}{Recompute or load from cache?}
}
\value{
A mwIPM model object.
}
\description{
Set site and recompute kernel.
}
\examples{
ipm \%<>\% setSite("BLD1")
}
