% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setSite}
\alias{setSite}
\title{Set site and recompute kernel.}
\usage{
setSite(obj, site, compute)
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{site}{Site ("Bertha", "BLD1", "BLD2", "GET", "GRN", "PWR", "SKY", "YTB")}

\item{compute}{Recompute or load from cache?}
}
\value{
A mwIPM model object.
}
\description{
Set site and recompute kernel.
}
\examples{
ipm \%<>\% setSite("BLD1")
}
