% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setVarDistMatrices.mwIPM}
\alias{setVarDistMatrices.mwIPM}
\title{Create herbivory matrix.}
\usage{
\method{setVarDistMatrices}{mwIPM}(obj, compute = FALSE,
  saveresults = FALSE, update = TRUE, perturb = rep(0, 7))
}
\arguments{
\item{obj}{A mwIPM model object.}

\item{update}{Update dependencies?}

\item{perturb}{Parameter perturbation vector for sensitivity analysis.}

\item{dist.herb}{Herbivory distribution.}

\item{distpars}{Determines in which scale (log or linear) perturbation is occurring.}
}
\value{
A mwIPM model object.
}
\description{
Create herbivory matrix.
}
