% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_ipm.R
\name{setVars.mwIPM}
\alias{setVars.mwIPM}
\title{Sets model variables.}
\usage{
\method{setVars}{mwIPM}(obj)
}
\arguments{
\item{obj}{A mwIPM model object.}
}
\value{
A mwIPM model object.
}
\description{
Sets model variables.
}
